lab02 p02 planina wrong answer test3

+++ lab-02/p02/main.cpp
@@ -10,7 +10,7 @@ int main()
 
     // int points = pow((iterations * 2 + 1), 2);
     int difference = 2;
-    int points = 3;
+    long long  points = 3;
     for (int i = 1; i < iterations; i++)
     {
         points += pow(difference, i);
